dj64: request dlmopen if asan is enabled
It turned out asan doesn't work with RTLD_DEEPBIND flag.
To avoid that flag, we need to request dlmopen().

+++ src/plugin/dj64/config/plugin_config.h
@@ -3,7 +3,7 @@
 
 struct pm_regs;
 struct djdev64_ops {
-    int (*open)(const char *path, unsigned flags);
+    int (*open)(const char *path, unsigned short flags);
     void (*close)(int handle);
     unsigned (*call)(int handle);
     unsigned (*ctrl)(int handle);
